server/requirements.txt
numpy~=2.4.3
fastapi~=0.135.1
pydantic~=2.12.5
opencv-python~=4.13.0.92
supabase>=2.10.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
xgboost>=2.0.0
google-generativeai>=0.3.0
ultralytics>=8.0.0
flask>=3.0.0
flask-cors>=4.0.0
scikit-learn>=1.3.0
matplotlib>=3.8.0